feat: mv mock dev db and model testing to test dir
